redesigned landing page as well as code clean up

+++ Android/app/src/main/java/com/fivegearszerochill/noted/room/entity/NoteCategoryEntity.java
@@ -22,7 +22,7 @@ public class NoteCategoryEntity implements CoreEntity {
         this.isActive = isActive;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(long categoryId) {
         this.categoryId = categoryId;
     }
